Add a box with specified color in an R Markdown file.

add_colored_box returns a box component generated by htmltools with specified color and styles.

Usage

add_colored_box(
  type = "blue-default",
  label = "",
  info = "place details here using info option",
  bgcolor = NULL,
  width = 0.5,
  halign = "c",
  top = FALSE
)

Arguments

type

One of:

NULL for no default color or label

'blue-default' for a steel-blue box

'gray-info' for a gray box

'blue-info' for a blue box

'green-reminder' for a green box

'yellow-warning' for a yellow box

'red-stop' for a red box

label

One of:

NULL for no label if type is NULL or using label set by type

A string shown on the top of box

info

A string including the main message of the box

bgcolor

NA or a length 3 vector with integer elements between 0 to 255

width

NA or a number between 0.25 to 0.95

halign

One of:

NA for center aligned

'c' for center aligned

'r' for right aligned

top

One of:

NA

FALSE for inline

TRUE for top-of-page

Value

A formatted htmltools tag object.

Examples

add_colored_box( type='blue-default', info='the document include information regarding...')
#> <div style=" background-color: rgba( 51,122,183 , 0.2); margin: 3px auto 3px auto; width: 50% ; border-width: 0px 0px 0px 3px; border-color: rgba( 51,122,183 ,1); border-style: solid; padding: 1px 1px 1px 0px; border-radius: 0px 0px 5px 0px;"><b> &nbsp; <span style='font-size:110%;'> &#9749; </span><i> Note </i></b> <div style="background-color: rgba(255,255,255,0.75); padding: 10px 20px 10px 20px; border-radius: 0px 0px 5px 0px;">the document include information regarding...</div></div>