Check two tables with unique matching ids and generate reports on duplicated ids and repeated columns when applicable..

bill_cross_check returns a merged data.table showing information regarding ids and repeated columns.

Usage

bill_cross_check(dt1 = NULL, dt2 = NULL, id = NULL, chk = NULL)

Arguments

dt1

A table.

dt2

A table.

id

A column name or a vector of column names

chk

A column name

Value

a data.table

Examples


bill_cross_check(data.frame(col1=c(1,2,3,3), col2=c('a','b','c','c'),
col3=c('-','=','+','-')),data.frame(col1=c(1,2,3), col2=c('a','b','c'),
col3=c('-','=','+')), id=c('col1','col2'), chk='col3')
#>    col1 col2 tbl_id:1 tbl_id:2  same
#> 1:    1    a        -        -  TRUE
#> 2:    2    b        =        =  TRUE
#> 3:    3    c      +;-        + FALSE